feat(models): addition of tflite models
This adds two tflite models for use with the project.

* bird_classifier_quantized.tflite: a basic object detection model
trained for the birds that hang out by my feeders

* bird_classifier_quantized_edgetpu.tflite: the same object detection
model compiled down for an edgeTPU to execute.

In addition, there is a file with the labels for the object detection
model in models/labels.txt

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>