feat(notebook): added Google Colab notebook
This adds a Google Colab specific notebook that runs all the scripts to
process the training data, train the model, quantize the model, and
export the models to TFLite and EdgeTPU optimized TFLite.

It has a few Colab specific elements like using the
`%tensorflow_version` magic and forms, these should be able to be easily
removed from the notebook if not working in Google Colab.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>